make safe produer testable

+++ Tests/Jolie/Safe/SafeProducer/messageForwarderService.ol
@@ -1,11 +1,13 @@
-include "database.iol"
-include "console.iol"
-include "time.iol"
+from database import Database, ConnectionInfo
+from console import Console
+from time import Time
 
 from .outboxService import KafkaOptions
 from .outboxService import PollSettings
 from .outboxService import StatusResponse
 from .simple-kafka-connector import SimpleKafkaConnector
+
+from ..test.producerTestTypes import MFSTestParams, TestException
 
 type ColumnSettings {
     .keyColumn: string
@@ -26,7 +28,8 @@ type ForwarderResponse{
 }
 
 interface MessageForwarderInterface {
-    RequestResponse: startReadingMessages ( ForwarderServiceInfo ) ( StatusResponse )
+    RequestResponse: setupTest( MFSTestParams )( bool )
+    OneWay: startReadingMessages ( ForwarderServiceInfo )
 }
 
 /**
@@ -39,6 +42,9 @@ service MessageForwarderService{
         Interfaces: MessageForwarderInterface
     }
     embed SimpleKafkaConnector as KafkaRelayer
+    embed Database as Database
+    embed Time as Time
+    embed Console as Console
 
     init {
         with ( connectionInfo ) 
@@ -53,30 +59,54 @@ service MessageForwarderService{
 
     /** Starts this service reading continually from the 'Messages' table */
     main{
-        [startReadingMessages( request )( response ) 
-        {
-            scope (ConnectToDatabase){      // Everything in this scope is simply to check that a connection to the database CAN be opened.
-                global.M_KafkaOptions << request.brokerOptions
-                response.status = 200
-                response.reason = "MessageForwarderService initialized sucessfully"
-                println@Console( "MessageForwarder connected to database!" )(  )
-            }
-
-        }] 
+        [startReadingMessages( request )] 
         {
             // Keep polling for messages at a given interval.
             while(true) {
-                connect@Database( connectionInfo )( void )      //I hate that i cannot keep the connection open from above, but likely scoping issue
+                scope (EnsureTestContinues){
+                    install( TestException =>  {
+                        println@Console("Exception thrown from MFS: " + main.TestException)()
+
+                        // Call this operation again
+                        scheduleTimeout@Time(500 {
+                            message << request
+                            operation = "startReadingMessages"
+                        })()
+                        
+                        // Rethrow the message such that tests can access it.
+                        throw (TestException, main.TestException)
+                    })
+                }
+
+                connect@Database( connectionInfo )( void )
                 query = "SELECT * FROM outbox LIMIT " + request.pollSettings.pollAmount
+                
+                // This throw should not cause desync issues
+                if (global.testParams.throw_before_check_for_messages){
+                    throw ( TestException, "throw_before_check_for_messages" )
+                }
+
                 query@Database(query)( pulledMessages )
                 println@Console( "Query '" + query + "' returned " + #pulledMessages.row + " rows " )(  )
+                
+                // This throw should not cause desync issues
+                if (global.testParams.throw_after_message_found){
+                    throw ( TestException, "throw_after_message_found" )
+                }
+                
                 if (#pulledMessages.row > 0){
                     for ( databaseMessage in pulledMessages.row ){
                         kafkaMessage.topic = "example"
                         kafkaMessage.key = databaseMessage.(request.columnSettings.keyColumn)
                         kafkaMessage.value = databaseMessage.(request.columnSettings.valueColumn)
-                        kafkaMessage.brokerOptions << global.M_KafkaOptions
+                        kafkaMessage.brokerOptions << request.brokerOptions
                         propagateMessage@KafkaRelayer( kafkaMessage )( kafkaResponse )
+
+                        // This throw will likely result in the message being sent twice!
+                        if (global.testParams.throw_after_send_but_before_delete){
+                            throw ( TestException, "throw_after_send_but_before_delete" )
+                        }
+
                         if (kafkaResponse.status == 200) {
                             update@Database( "DELETE FROM outbox WHERE " + ( request.columnSettings.idColumn ) + " = " + databaseMessage.(request.columnSettings.idColumn) )( updateResponse )
                         }
@@ -85,5 +115,10 @@ service MessageForwarderService{
                 sleep@Time( request.pollSettings.pollDurationMS )(  )
             }
         }
+        
+        [setupTest( request )( response ){
+            global.testParams << request
+            response = true
+        }]
     }
 }
